% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{compute_eigengene_matrix}
\alias{compute_eigengene_matrix}
\title{Compute community signatures (eigengenes)}
\usage{
compute_eigengene_matrix(ex, membership_matrix, cutoff = 5, pc_flag = TRUE)
}
\arguments{
\item{ex}{matrix of bulk RNA-seq or microarray gene expression data.
This should be in log space and greater than or equal to 0.}

\item{membership_matrix}{a community membership (kME) matrix with genes as
rows and communities as columns. Often \code{community_membership} or
\code{full_community_membership} output from \code{\link[=icwgcna]{icwgcna()}}}

\item{cutoff}{number of top genes to use when computing community signatures}

\item{pc_flag}{indicator. TRUE (default) means to use the 1st principal component
(corrected for direction). FALSE uses the mean of scaled and centered top genes.}
}
\value{
A matrix with rows being the community signature and columns being samples
}
\description{
Computes the community signatures (eigengenes) for an expression matrix
given a particular community membership (kME) matrix
}
\details{
This can be used to compute community signatures in a new expression dataset.
Note, community signatures are not corrected by \code{\link[=icwgcna]{icwgcna()}}
iterations so it will not match signatures output from \code{\link[=icwgcna]{icwgcna()}}
if it is run on the network construction dataset.

When using these community signatures for modeling it may be best to include
interaction terms or use tree based
methods since dependencies are not addressed in this output matrix.
}
\examples{
\dontrun{
library("UCSCXenaTools")
luad <- getTCGAdata(project = "LUAD", mRNASeq = TRUE, mRNASeqType = "normalized",
clinical = FALSE, download = TRUE)
ex <- as.matrix(data.table::fread(luad$destfiles), rownames = 1)

results <- icwgcna(ex)

compute_eigengene_matrix(ex, results$community_membership)
}

}
