% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_panglaoDB_enrichment}
\alias{compute_panglaoDB_enrichment}
\title{Compute Cell Type Enrichments Using panglaoDB Cell Markers}
\usage{
compute_panglaoDB_enrichment(
  t_memb,
  K = 100,
  memb_cut = 0.65,
  pangDB = data.table::fread(pangDB_link),
  prolif = prolif_names
)
}
\arguments{
\item{t_memb}{\code{community_membership} or \code{full_community_membership} values from \code{\link[=icwgcna]{icwgcna()}}}

\item{K}{cutoff for ???}

\item{memb_cut}{cutoff for ???}

\item{pangDB}{panglaoDB cell markers database. Default is to read data from the url \link{pangDB_link}}

\item{prolif}{list of proliferation genes to check. Default is to use \link{prolif_names}}
}
\value{
Returns a list with the following items:
\itemize{
\item \code{top_enr} - ???
\item \code{full_enr} - ???
}
}
\description{
compute cell type enrichments using \href{https://academic.oup.com/database/article/doi/10.1093/database/baz046/5427041?login=true}{panglaoDB cell markers}
}
\examples{

\dontrun{
pangDB <- data.table::fread(pangDB_link)
compute_panglaoDB_enrichment()}

}
