% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{compute_xCell_enrichment}
\alias{compute_xCell_enrichment}
\title{Compute Cell Type Enrichments Using xCell Cell Markers}
\usage{
compute_xCell_enrichment(
  membership_matrix,
  K = 100,
  memb_cut = 0.65,
  p_cut = 0.001
)
}
\arguments{
\item{membership_matrix}{a community membership (kME) matrix with genes as
rows and communities as columns. Often \code{community_membership} or
\code{full_community_membership} output from \code{\link[=icwgcna]{icwgcna()}}}

\item{K}{cutoff for top community genes to include for computing enrichment.
Used in an AND condition with \code{memb_cut}.}

\item{memb_cut}{cutoff as a membership score threshold for determining top
community genes for computing enrichment. Used in an AND condition with \code{K}.}

\item{p_cut}{p value cutoff for determining importance. If all p values are
below \code{p_cut} for a community, no cell type is selected}
}
\value{
Returns a list with the following items:
\itemize{
\item \code{top_enr} - the most significant cell type from the enrichment scores.
\item \code{full_enr} - all panglaoDB cell type enrichment scores for all communities.
}
}
\description{
Compute cell type enrichments using
\href{https://github.com/dviraran/xCell}{xCell cell markers} with Fisher test.
}
\details{
note that this is distinct from running xCell which is run on expression data. This
is an enrichment using their cell markers.
}
\examples{
\dontrun{
library("UCSCXenaTools")
luad <- getTCGAdata(
  project = "LUAD", mRNASeq = TRUE, mRNASeqType = "normalized",
  clinical = FALSE, download = TRUE
)
ex <- as.matrix(data.table::fread(luad$destfiles), rownames = 1)
results <- icwgcna(ex)

compute_xCell_enrichment(results$community_membership)
}

}
