% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{display_top_genes}
\alias{display_top_genes}
\title{Identify Top Genes and Values of Communities}
\usage{
display_top_genes(
  membership_matrix,
  K = 10,
  output = c("both", "genes", "values")
)
}
\arguments{
\item{membership_matrix}{a community membership (kME) matrix with genes as
rows and communities as columns. Often \code{community_membership} or
\code{full_community_membership} output from \code{\link[=icwgcna]{icwgcna()}}}

\item{K}{number of genes to find.}

\item{output}{Should output show both top genes and values, genes only, or
values only}
}
\value{
K x Communities data.frame of top genes and ranks (if \code{output} = "genes")

K x Communities data.frame of top values and ranks (if \code{output} = "values")

Long data.frame with columns Community, Gene, Value, Rank
(if \code{output} = "both")
}
\description{
Identify Top Genes and Values of Communities
}
\examples{
\dontrun{
library("UCSCXenaTools")
luad <- getTCGAdata(
  project = "LUAD", mRNASeq = TRUE, mRNASeqType = "normalized",
  clinical = FALSE, download = TRUE
)
ex <- as.matrix(data.table::fread(luad$destfiles), rownames = 1)
results <- icwgcna(ex)

top_genes <- display_top_genes(
    results$community_membership,
    K = 100,
    output = "both")
}
}
