% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenMapMatMult}
\alias{eigenMapMatMult}
\title{Faster Matrix Multiplication}
\usage{
eigenMapMatMult(A, B)
}
\arguments{
\item{A, B}{Matrices.}
}
\description{
Faster matrix multiplication using C++.
}
