% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fastTOMwrapper}
\alias{fastTOMwrapper}
\title{fastTOMwrapper}
\usage{
fastTOMwrapper(
  X,
  expo = 6,
  Method = c("pearson", "spearman"),
  mat_mult_method = c("Rfast", "RcppEigen")
)
}
\arguments{
\item{X}{a gene expression matrix w each column being one sample and N rows representing genes. X should be in log space (usually between 0 and 20)}

\item{expo}{the power to raise the similarity measure to default = 6. If set to NULL, angular distance is used to applied to the similarity measure ( asin(x) / (pi/2) ).}

\item{Method}{"pearson" or "spearman" the similarty measure to use}

\item{mat_mult_method}{method for large matrix multiplication, "Rfast" (default) or "RcppEigen" (see \code{details} in \code{icwgcna()})}
}
\value{
A N x N distance matrix with smaller values indicating more related genes.
}
\description{
Topological overlap map (TOM, from \href{https://pubmed.ncbi.nlm.nih.gov/12202830/}{Ravasz, et al (2002). Science}) wrapper using the \href{https://cran.r-project.org/web/packages/Rfast/}{Rfast package}
to speed up calculations. This function can compute TOM for 24K gene matrix in 8 minute of an AWS-EC2 c5.18xlarge instance, though in practice we run it on subsets of most variable genes.
Adjacenty measur options are Pearson or Spearman correlation raised to a power and angular distance. Note that we only use signed and weighted adjacencies. If users are interested in genes negatively associated
with a community they should check community memberships (kME) output from the main function icwgcna().
}
