% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{find_unique_top_genes}
\alias{find_unique_top_genes}
\title{Identify Top Gene of Communities that are unique (only belong to one community)}
\usage{
find_unique_top_genes(membership_matrix, K = 10, maxIt = 10)
}
\arguments{
\item{membership_matrix}{a community membership (kME) matrix with genes as
rows and communities as columns. Often \code{community_membership} or
\code{full_community_membership} output from \code{\link[=icwgcna]{icwgcna()}}}

\item{K}{number of unique genes to find.}

\item{maxIt}{maximum number of iterations to avoid looping to meaningless top genes}
}
\value{
data.frame with key = community and value = top gene ID. Top genes will only
be associated with one module.
}
\description{
Identify Top Gene of Communities that are unique (only belong to one community)
}
\examples{
\dontrun{
library("UCSCXenaTools")
luad <- getTCGAdata(
  project = "LUAD", mRNASeq = TRUE, mRNASeqType = "normalized",
  clinical = FALSE, download = TRUE
)
ex <- as.matrix(data.table::fread(luad$destfiles), rownames = 1)
results <- icwgcna(ex)

unique_top_genes <- find_unique_top_genes(
    results$community_membership,
    K = 100,
    maxIt = 10)
}

}
