% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icWGCNA-package.R
\docType{package}
\name{icWGCNA-package}
\alias{icWGCNA}
\alias{icWGCNA-package}
\title{icWGCNA: Iterative Correcting Weighted Gene Co-expression Network Analysis}
\description{
Iterative Correcting Weighted Gene Co-expression Network Analysis function for constructing a gene network from a gene expression matrix. The algorithm: 1. Constructs a signed wgcna network. 2. Drops correlated modules based on kurtosis. 3. Regresses out the largest community from the expression data. 4. Repeats steps 1-3 until a maximum number of communities or iterations is reached.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Systems-Methods/icWGCNA}
  \item Report bugs at \url{https://github.com/Systems-Methods/icWGCNA/issues}
}

}
\author{
\strong{Maintainer}: Mike Mason \email{mike.mason@bms.com} (\href{https://orcid.org/0000-0002-5652-7739}{ORCID})

Authors:
\itemize{
  \item Jimmy Fulp \email{william.fulp@bms.com} (\href{https://orcid.org/0000-0001-6888-0794}{ORCID})
}

Other contributors:
\itemize{
  \item Alexandre Alloy \email{alexandre.alloy@bms.com} (\href{https://orcid.org/0000-0001-8280-0543}{ORCID}) [contributor]
}

}
\keyword{network-analysis}
