% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icwgcna.R
\name{icwgcna}
\alias{icwgcna}
\title{icwgcna Iterative Correcting Weighted Gene Co-expression Network Analysis function constructing a network from an expresion matrix.}
\usage{
icwgcna(
  ex,
  expo = 6,
  Method = "pearson",
  q = 0.5,
  maxIt = 25,
  maxComm = 100,
  corCut = 0.6
)
}
\arguments{
\item{ex}{matrix of bulk RNA-seq or microarray gene expresion data}

\item{expo}{exponted to use for soft thresholding}

\item{Method}{correlation to use for distance measure, "pearson" (default) or "spearman"}

\item{q}{quantile (0-1) for first round filtering based on mean expression and standard deviation}

\item{maxIt}{maximum number of iterations must be 25 or less}

\item{maxComm}{maximum number of communities to be found}

\item{corCut}{correlation threshold used for dropping communities}
}
\value{

}
\description{
icwgcna Iterative Correcting Weighted Gene Co-expression Network Analysis function constructing a network from an expresion matrix.
}
