% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{pangDB_link}
\alias{pangDB_link}
\title{pangDB_link}
\format{
An object of class \code{character} of length 1.
}
\usage{
pangDB_link
}
\description{
Current link used to connect to panglaoDB cell markers
}
\keyword{datasets}
