% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\docType{data}
\name{prolif_names}
\alias{prolif_names}
\title{Current list of proliferation genes to check}
\format{
An object of class \code{character} of length 27.
}
\usage{
prolif_names
}
\description{
Current list of proliferation genes to check
}
\examples{

prolif_names

}
\keyword{datasets}
