% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{compute_panglaoDB_enrichment}
\alias{compute_panglaoDB_enrichment}
\title{Compute Cell Type Enrichments Using panglaoDB Cell Markers}
\usage{
compute_panglaoDB_enrichment(
  membership_matrix,
  K = 100,
  memb_cut = 0.65,
  pangDB = data.table::fread(pangDB_link),
  prolif = prolif_names
)
}
\arguments{
\item{membership_matrix}{a community membership (kME) matrix with genes as
rows and communities as columns. Often \code{community_membership} or
\code{full_community_membership} output from \code{\link[=icwgcna]{icwgcna()}}}

\item{K}{cutoff for top community genes to include for computing enrichment.
Used in an AND condition with memb_cut.}

\item{memb_cut}{cutoff as a membership score threshold for determining top
community genes for computing enrichment.  Used in an AND condition with K.}

\item{pangDB}{panglaoDB cell markers database. Default is to read data from t
he url \link{pangDB_link}}

\item{prolif}{list of proliferation genes to check.
Default is to use \link{prolif_names}}
}
\value{
Returns a list with the following items:
\itemize{
\item \code{top_enr} - the most significant cell type from the enrichment scores.
\item \code{full_enr} - all panglaoDB cell type enrichment scores for all communities.
}
}
\description{
compute cell type enrichments using \href{https://academic.oup.com/database/article/doi/10.1093/database/baz046/5427041?login=true}{panglaoDB cell markers}
using Fisher test.
}
\examples{
\dontrun{
library("UCSCXenaTools")
luad <- getTCGAdata(
  project = "LUAD", mRNASeq = TRUE, mRNASeqType = "normalized",
  clinical = FALSE, download = TRUE
)
ex <- as.matrix(data.table::fread(luad$destfiles), rownames = 1)
results <- icwgcna(ex)

pangDB <- data.table::fread(pangDB_link)
compute_panglaoDB_enrichment(results$community_membership, pangDB = pangDB)
}

}
