% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icwgcna.R
\name{icwgcna}
\alias{icwgcna}
\title{icwgcna}
\usage{
icwgcna(
  ex,
  expo = 6,
  Method = c("pearson", "spearman"),
  q = 0.3,
  maxIt = 10,
  maxComm = 100,
  corCut = 0.8,
  covCut = 0.33,
  mat_mult_method = c("Rfast", "penppml")
)
}
\arguments{
\item{ex}{matrix of bulk RNA-seq or microarray gene expression data}

\item{expo}{exponted to use for soft thresholding}

\item{Method}{correlation to use for distance measure, "pearson" (default) or "spearman"}

\item{q}{quantile (0-1) for first round filtering based on mean expression and standard deviation}

\item{maxIt}{maximum number of iterations must be 25 or less}

\item{maxComm}{maximum number of communities to be found}

\item{corCut}{correlation threshold used for dropping communities}

\item{covCut}{coeficient of variation (CoV) quantile threshold to use at each iteration  for selecting genes to build network. covCut = .667 would use the top third of genes based on CoV after regressing out largest community}

\item{mat_mult_method}{method for large matrix multiplication, "Rfast" (default) or "penppml" (see \code{details})}
}
\value{
Returns a list with the following items:
\itemize{
\item \code{community_membership} -
\item \code{community_signature} -
\item \code{full_community_membership} -
\item \code{full_community_signature} -
\item \code{controlled_for} -
}
}
\description{
Iterative Correcting Weighted Gene Co-expression Network Analysis function constructing a network from an expression matrix.
}
\details{
Iterative Correcting Weighted Gene Co-expression Network Analysis function for constructing a gene network from a gene expression matrix. The algorithm:
\enumerate{
\item Constructs a signed wgcna network
\item Drops correlated modules based on kurtosis.
\item Regresses out the largest community from the expression data.
\item Repeats steps 1-3 until a maximum number of communities or iterations is reached.
}

Some differences from standard WGNCA (Horvath/Langfelder)
\itemize{
\item Makes heavy use of \href{https://cran.r-project.org/web/packages/Rfast/}{Rfast package} to compute adjacencies and TOM to enable iterative network creation on > 20K features.
\item Uses signed adjacency in order to avoid possible distortions of community signatures (eigengenes).
\item Iteratively regresses out strongest community in order to facilitate discovery of communities possibly obscured larger module(s).
\item Clustering does not focus on merging communities but dropping to identify strongest module(s).
\item Enables Spearman correlation for constructing adjacency matrix instead of Pearson to enable robust application in RNA-seq and micro-array data. Future updates may include mutual information
}

For matrix multiplication the option "Rfast" will use \code{\link[Rfast:mat.mult]{Rfast::mat.mult()}},
which takes advantage of parallel processing across multiple cores. The option
"penppml" will use \code{\link[penppml::eigenMapMatMult]{penppml:::eigenMapMatMult()}},
which tends to be faster when using
a single core, but does not take advantage of parallel processing across
multiple cores. If running this on a cluster with access to many computer core
there is a significant performance advantage to using \code{\link[Rfast:mat.mult]{Rfast::mat.mult()}}
}
