% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{map_eigengenes_on_seurat}
\alias{map_eigengenes_on_seurat}
\title{Map Eigengenes on a Seurat Object}
\usage{
map_eigengenes_on_seurat(
  seurat_obj,
  membership,
  cutoff_method = c("value", "top_gene", "both"),
  value_cutoff = 0.75,
  top_genes_cutoff = 10,
  assay = "RNA",
  slot = "data",
  prefix = NULL,
  suffix = "_UCell",
  ncores = 1
)
}
\arguments{
\item{seurat_obj}{Seurat Object}

\item{membership}{a data.frame or matrix of continuous values, with genes as
rows and communities as columns. Often this is the \code{community_membership} or
\code{full_community_membership} output from \code{\link[=icwgcna]{icwgcna()}}.}

\item{cutoff_method}{should cutoff be based on a value, number of
top genes, or both method. Both method will apply the top gene method,
but only for genes higher than the \code{value_cutoff}.}

\item{value_cutoff}{value cutoff (ignored when \code{cutoff_method} = "top_gene")}

\item{top_genes_cutoff}{number of top genes to include
(ignored when \code{cutoff_method} = "value")}

\item{assay}{Seurat object assay element to use}

\item{slot}{Pull out data from this slot of the Seurat object}

\item{prefix}{prefix to add to column names of the Seurat object meta.data}

\item{suffix}{suffix to add to column names of the Seurat object meta.data}

\item{ncores}{Number of processors to parallelize computation for
\code{\link[UCell:AddModuleScore_UCell]{UCell::AddModuleScore_UCell()}}}
}
\value{
Seurat Object with additional meta.data columns of community results
}
\description{
Use scores to calculate module scores for feature expression
programs in single cells and applies to Seurat object using
\code{\link[UCell:AddModuleScore_UCell]{UCell::AddModuleScore_UCell()}}
}
\details{
For typical Seurat objects this function may take a few minutes if only
using one core.
}
\examples{
\dontrun{
library("UCSCXenaTools")
luad <- getTCGAdata(
  project = "LUAD", mRNASeq = TRUE, mRNASeqType = "normalized",
  clinical = FALSE, download = TRUE
)
ex <- as.matrix(data.table::fread(luad$destfiles), rownames = 1)
results <- icwgcna(ex)

unique_top_genes <- map_eigengenes_on_seurat(
    SeuratObject::pbmc_small
    results$community_membership
    )
}
}
