% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\docType{data}
\name{pangDB_link}
\alias{pangDB_link}
\title{Current link used to connect to \href{https://academic.oup.com/database/article/doi/10.1093/database/baz046/5427041?login=true}{panglaoDB cell markers}}
\format{
An object of class \code{character} of length 1.
}
\usage{
pangDB_link
}
\description{
Current link used to connect to \href{https://academic.oup.com/database/article/doi/10.1093/database/baz046/5427041?login=true}{panglaoDB cell markers}
}
\details{
pangDB_link = \url{https://panglaodb.se/markers/PanglaoDB_markers_27_Mar_2020.tsv.gz}
}
\examples{

pangDB_link

}
\keyword{datasets}
